Changed sample data type from f32 to i16 and set maximum polyphony to 100000. Removed fade-in.

+++ crates/ksynth-core/src/sample.rs
@@ -1,10 +1,10 @@
 #[derive(Clone, Debug)]
 pub struct Sample {
-    pub sample_data: Vec<f32>,
+    pub sample_data: Vec<i16>,
 }
 
 impl Sample {
-    pub fn new(data: Vec<f32>) -> Self {
+    pub fn new(data: Vec<i16>) -> Self {
         Self { sample_data: data }
     }
